Restructure buildSrc setup to be more generic
This is some groundwork to add future tasks that aren't related to kotlin generation
