--- buildSrc/src/main/resources/META-INF/gradle-plugins/com.jakewharton.kotlingen.properties
@@ -1,1 +1,0 @@
-implementation-class=com.jakewharton.kotlingen.KotlinGenPlugin
