+++ buildSrc/src/main/resources/META-INF/gradle-plugins/rxbinding-module.properties
@@ -1,0 +1,1 @@
+implementation-class=com.jakewharton.rxbindingproject.RxBindingPlugin
